feat(useSubPagesDataForCurrPath): Extract `useSubPagesDataForCurrPath` hook
