+++ pages/exercises/[exercise].tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import dynamic from "next/dynamic";
 import CircularProgress from "@mui/material/CircularProgress";
 
@@ -6,30 +6,12 @@ import type { NextPage } from "next";
 
 import { sectionConfigs } from "@/src/metadata/sectionConfigs";
 import PageSection from "@/src/components/page/PageSection";
-import SectionMetadata from "@/src/model/SectionMetadata";
-import { useAppGlobalStateContext } from "@/src/contexts/AppGlobalStateContext";
+import { useSubPagesDataForCurrPath } from "@/src/hooks/useSubPagesDataForCurrPath";
 
 const Exercise = () => {
-    const { currDisplayPath } = useAppGlobalStateContext();
+    const { data: pageData, loading } = useSubPagesDataForCurrPath(sectionConfigs.exercises);
 
-    const pagePath = currDisplayPath.split("/").at(-2);
-
-    const [isLoading, setLoading] = useState(true);
-    const [pageData, setPageData] = useState<SectionMetadata | undefined>(undefined);
-
-    useEffect(() => {
-        sectionConfigs.exercises.getSubPages().then(({ data }) => {
-            setLoading(false);
-
-            if (!pagePath || !data || !data[pagePath]) {
-                return;
-            }
-
-            setPageData(data[pagePath]);
-        });
-    }, [pagePath]);
-
-    if (isLoading) {
+    if (loading) {
         return <CircularProgress />;
     }
     if (pageData) {
